Fix Clang waring.

+++ src/BasisSet/Atom/kappa/Imp/Slater_BF.C
@@ -13,7 +13,7 @@ namespace Atom_kappa
 namespace Slater
 {
 
-Large_BasisFunction::Large_BasisFunction(double ex, int _kappa, int _mj, double norm)
+Large_BasisFunction::Large_BasisFunction(double ex, int _kappa, double _mj, double norm)
     : itsExponent     (ex)
     , kappa(_kappa), mj(_mj), l(Omega_kmj_Sym::l(kappa))
     , itsNormalization(norm)
